[graphite] Use TextureFormat in AttachmentDesc
This CL covers a fair amount of ground as TextureInfo extended fairly
far throughout the code base. So from a top-down overview of the
changes:

- RenderPassDesc's AttachmentDesc now stores a TextureFormat, load op, store op, and sample count. It no longer stores TextureInfo.
   - This means RenderPassDescs no longer depend on backend indirection, so pipeline keys + labels, precompilation data, and textureinfo serialization logic is simplified (or removed in the case of TextureInfo serialization).
   - Precompilation serialized keys no longer have the non-key data in a RenderPassDesc (e.g. clear color).
   - Pipeline labels are made more compact (see changes to ChromePrecompileTest)
   - These changes propagated through GlobalCache, PrecompileContext, and SerializationUtils.
   - To help with data sizes (AttachmentDesc being 32-bits in particular), sample count variables are now typed to be uint8_ts in most places.

- RenderPassDesc::Make doesn't need to query Caps for TextureInfos anymore, so the Caps APIs are updated to instead return more specific pieces of data: DepthStencilFlags -> TextureFormat, supported sample count.
    - These are all implemented very similarly between the different subclasses, so eventually format tables can be stored in the base Caps and then the implementations can just query the table w/o needing to know the specific backend format value.
    - ResourceProvider and Caps don't need to have separate functions for creating Depth/stencil or MSAA attachments. Caps instead gets a default attachment texture info that can be used for both cases, and ResourceProvider switches to findOrCreateShareableTexture.
    - This requires some code changes in RenderPassTask
    - Format tables also now consider depth/stencil formats to be "renderable" so that the general `Caps::isRenderable(TextureInfo)` returns true for them, which then simplifies some logic in Texture factories.

- Every Caps subclass migrated their old TextureInfo query functions to the new APIs, and updated their pipeline key functions.
   - Most backends had already declared internal functions similar to the new APIs to check for sample count support or pick the depth/stencil format, so they just moved to the new entry point.
   - Pipeline key generation and extracting back to a RenderPassDesc affected a fair amount of code but was mostly straight forward.

- The backends had had BackendFormatToTextureFormat functions. Each now also has a TextureFormatToBackendFormat function. They are both generated using a macro-of-macros so that there's one bidirectional table defined for each backend.

- VulkanRenderPass::Metadata was able to be simplified more because it could rewrite AttachmentDescs in place to handle compatibility. It was also pretty straight forward to update it to track attachment indices, so now VulkanRenderPasses can be created from just the Metadata instead of the RenderPassDesc.
   - The msaa loading pipelines are now cached by their Metadata instead of going through a GraphiteResourceKey, so there's fewer entry points to creating/looking-up render passes now.

Bug: b/390473370
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/963157
Commit-Queue: Michael Ludwig <michaelludwig@google.com>

+++ include/gpu/graphite/ContextOptions.h
@@ -45,7 +45,7 @@ struct SK_API ContextOptions {
      *
      * If <= 1, Graphite will disable internal code paths that use multisampling.
      */
-    int fInternalMultisampleCount = 4;
+    uint8_t fInternalMultisampleCount = 4;
 
     /**
      * Will the client make sure to only ever be executing one thread that uses the Context and all
